complteted assignment Student File

+++ src/ftd_8_10/day3.java
@@ -1,26 +1,35 @@
 package ftd_8_10;
-
-import java.io.BufferedReader;
 import java.io.BufferedWriter;
 import java.io.File;
-import java.io.FileReader;
+
 import java.io.FileWriter;
+import java.util.Scanner;
+
+import javax.swing.JFileChooser;
 
 public class day3 {
-
+	final static JFileChooser fc = new JFileChooser();
 	public static void main(String[] args) throws Exception {
-		File file = new File("text.txt");
-		BufferedReader read = new BufferedReader(new FileReader(file));
-		BufferedWriter write = new BufferedWriter(new FileWriter(file));
-		while(true){
-			write.write(read.readLine());
-			
-
-			
-			
+		fc.showOpenDialog(null);
+		File file = fc.getSelectedFile();	
+		if(!file.exists()){
+			file.createNewFile();
+			Student stu = new Student("stu","Tennesse","Memphis");
+			BufferedWriter write = new BufferedWriter(new FileWriter(file));
+			write.write(stu.toString());
+			write.close();
+		}
+		else{
+			Scanner sc= new Scanner(file);
+			while(sc.hasNext()){
+				Student stu =new Student(sc.next(),sc.next(),sc.next());
+				System.out.println(stu.toString());
+			}
+			sc.close();
 			
 			
 		}
+		
 		}
 	}
